[AM][OPS-9321] move IncomeInput to ssttpaffordability package
